:art: move register related functions from manager.py to publish.py
