+++ example/Test_torchfamily.yml
@@ -1,4 +1,8 @@
-weight: "~/.modelci/Alexnext/pytorch-pytorch/image_classification/1.pth"
+weight: "~/.modelci/Test_torchfamily/pytorch-pytorch/image_classification/1.pth"
+architecture: Alexnet
+framework: PYTORCH
+engine: PYTORCH
+version: 1
 dataset: ImageNet
 task: IMAGE_CLASSIFICATION
 metric:
