+++ modelci/hub/manager.py
@@ -16,136 +16,19 @@ import os
 import subprocess
 
 from pathlib import Path
-from shutil import copy2, make_archive
-from typing import Union, List
 
-import cv2
-import tensorflow as tf
-import yaml
-
-from modelci.hub.client.onnx_client import CVONNXClient
-from modelci.hub.client.tfs_client import CVTFSClient
-from modelci.hub.client.torch_client import CVTorchClient
-from modelci.hub.client.trt_client import CVTRTClient
+from typing import List
 from modelci.hub import converter
 
-from modelci.hub.utils import TensorRTPlatform, parse_path_plain, generate_path_plain
+from modelci.hub.utils import TensorRTPlatform
 from modelci.persistence.service import ModelService
-from modelci.persistence.service_ import save
+
 from modelci.types.bo import Task, ModelVersion, Framework, ModelBO
 
-__all__ = ['get_remote_model_weight', 'register_model', 'register_model_from_yaml', 'retrieve_model',
+__all__ = ['get_remote_model_weight', 'retrieve_model',
            'retrieve_model_by_task', 'retrieve_model_by_parent_id']
 
-from modelci.types.models.common import Engine, ModelStatus
-
-from modelci.types.models import MLModelFromYaml, MLModel
-
-
-def register_model(
-        model: MLModel,
-        convert: bool = True,
-        profile: bool = True,
-) -> List[MLModel]:
-    """Upload a model to ModelDB.
-    This function will upload the given model into the database with some variation. It may optionally generate a
-        branch of models (i.e. model family) with different optimization techniques. Besides, a benchmark will be
-        scheduled for each generated model, in order to gain profiling results for model selection strategies.
-        In the `no_generate` model(i.e. `no_generate` flag is set to be `True`), `architecture`, `framework`, `engine`
-        and `version` could be None. If any of the above arguments is `None`, all of them will be auto induced
-        from the origin_model path. An `ValueError` will be raised if the mata info cannot be induced.
-
-    TODO:
-        This function has a super comprehensive logic, need to be simplified.
-
-    Arguments:
-        model: Required inputs for register a model. All information is wrapped in such model.
-        convert (bool): Flag for generation of model family. Default to True.
-        profile (bool): Flag for profiling uploaded (including converted) models. Default to True.
-    """
-    models = list()
-
-    model_dir_list = list()
-    model.model_status = [ModelStatus.PUBLISHED]
-    models.append(save(model))
-
-    # generate model family
-    if convert:
-        model_dir_list.extend(converter.generate_model_family(model))
-
-    # register
-    model_data = model.dict(exclude={'weight', 'id', 'model_status', 'engine'})
-    for model_dir in model_dir_list:
-        parse_result = parse_path_plain(model_dir)
-        engine = parse_result['engine']
-
-        model_cvt = MLModel(**model_data, weight=model_dir, engine=engine, model_status=[ModelStatus.CONVERTED])
-        models.append(save(model_cvt))
-
-    # profile registered model
-    if profile:
-        from modelci.controller import job_executor
-        from modelci.controller.executor import Job
-
-        file = tf.keras.utils.get_file(
-            "grace_hopper.jpg",
-            "https://storage.googleapis.com/download.tensorflow.org/example_images/grace_hopper.jpg")
-        test_img_bytes = cv2.imread(file)
-
-        kwargs = {
-            'repeat_data': test_img_bytes,
-            'batch_size': 32,
-            'batch_num': 100,
-            'asynchronous': False,
-        }
-
-        for model in models:
-            model.model_status = [ModelStatus.PROFILING]
-            ModelService.update_model(model)
-            kwargs['model_info'] = model
-            engine = model.engine
-
-            if engine == Engine.TORCHSCRIPT:
-                client = CVTorchClient(**kwargs)
-            elif engine == Engine.TFS:
-                client = CVTFSClient(**kwargs)
-            elif engine == Engine.ONNX:
-                client = CVONNXClient(**kwargs)
-            elif engine == Engine.TRT:
-                client = CVTRTClient(**kwargs)
-            else:
-                raise ValueError(f'No such serving engine: {engine}')
-
-            job_cuda = Job(client=client, device='cuda:0', model_info=model)
-            # job_cpu = Job(client=client, device='cpu', model_info=model)
-            job_executor.submit(job_cuda)
-            # job_executor.submit(job_cpu)
-
-    return models
-
-
-def register_model_from_yaml(file_path: Union[Path, str]):
-    # check if file exist
-    file_path = Path(file_path)
-    assert file_path.exists(), f'Model definition file at {str(file_path)} does not exist'
-
-    # read yaml
-    with open(file_path) as f:
-        model_config = yaml.safe_load(f)
-    model_yaml = MLModelFromYaml.parse_obj(model_config)
-    # copy model weight to cache directory
-    model_in_saved_path = model_yaml.saved_path
-    if model_in_saved_path != model_yaml.weight:
-        copy2(model_yaml.weight, model_in_saved_path)
-
-    # zip weight folder
-    if model_yaml.engine == Engine.TFS:
-        weight_dir = model_yaml.weight
-        make_archive(weight_dir.with_suffix('.zip'), 'zip', weight_dir)
-
-    model_data = model_yaml.dict(exclude_none=True, exclude={'convert', 'profile'})
-    model = MLModel.parse_obj(model_data)
-    register_model(model, convert=model_yaml.convert, profile=model_yaml.profile)
+from modelci.types.models.common import Engine
 
 
 def get_remote_model_weight(model: ModelBO):
